COMP: Remove warnings and fix shared build

+++ Utilities/vtkexodus2/excn2s.c
@@ -350,7 +350,7 @@ int ex_cvt_nodes_to_sides(int exoid,
     elem_blk_parms[i].num_nodes_per_elem = num_nodes_per_elem;
     elem_blk_parms[i].num_attr = num_attr;
 
-    for (m=0; m < strlen(elem_type); m++)
+    for (m=0; (size_t)m < strlen(elem_type); m++)
       elem_blk_parms[i].elem_type[m] = 
               toupper((int)elem_type[m]);
     elem_blk_parms[i].elem_type[m] = '\0';
