+++ Utilities/vtknetcdf/dim.c
@@ -99,7 +99,7 @@ find_NC_Udim(const NC_dimarray *ncap, NC_dim **dimpp)
   {
     /*EMPTY*/
   }
-  if(dimid >= ncap->nelems)
+  if((size_t)dimid >= ncap->nelems)
     return(-1); /* not found */
   /* else, normal return */
   if(dimpp != NULL)
@@ -137,7 +137,7 @@ NC_finddim(const NC_dimarray *ncap, const char *name, NC_dim **dimpp)
   {
     /*EMPTY*/
   }
-  if(dimid >= ncap->nelems)
+  if((size_t)dimid >= ncap->nelems)
     return(-1); /* not found */
   /* else, normal return */
   if(dimpp != NULL)
